changed query to send user name instead of id
